Facet title renaming, autosearch using timeout.
Add a mapping from internal facet field names to something that may be
more helpful to the user.

Change the implementation of autosearch to use a timeout, so that
autosearch happens a configurable amount of time after the last keyup
event in the input box.

Fixed a bug relating to the "close" button in the filter list.

Increased the width of the facet panel.

Refactoring: compilation of the handlebars templates can be done once.